requirements.txt
chardet==4.0.0
charset_normalizer==3.4.3
concurrent_log_handler==0.9.19
Flask==3.1.2
gunicorn==20.1.0
httpx==0.28.1
langchain==1.0.2
langchain_community==0.4.1
langchain_core==1.0.1
langchain_openai==1.0.1
langchain_aws==1.0.0
langchain-text-splitters==1.0.0
langchain_qdrant==1.1.0
langfuse==3.8.1
langgraph==1.0.1
mkdocs==1.5.3
mkdocs-material==9.5.18
mkdocs-material-extensions==1.3.1
mkdocs-macros-plugin==1.4.1
nest_asyncio==1.6.0
numpy==2.3.2
openpyxl==3.0.7
pydantic==2.12.3
pyotp==2.6.0
python-dotenv==1.2.1
pyyaml_env_tag==0.1
pymdown-extensions==10.16.1
qdrant_client==1.15.1
qrcode==7.4.2
Requests==2.32.5
sentence_transformers==5.1.0
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
